@@ -1,14 +1,27 @@
+import polyglotI18nProvider from 'ra-i18n-polyglot';
 import {
     Admin,
-    LoginWithEmail,
+    CustomRoutes,
     Resource,
-    ListGuesser,
-    EditGuesser,
+    localStorageStore,
+    useStore,
+    StoreContextProvider,
 } from 'react-admin';
+import { Route } from 'react-router';
 import { Client } from 'appwrite';
 import { appWriteDataProvider, appWriteAuthProvider } from 'ra-appwrite';
 
-import contacts from './contacts';
+import categories from './categories';
+import { Dashboard } from './dashboard';
+import englishMessages from './i18n/en';
+import invoices from './invoices';
+import { Layout, Login } from './layout';
+import orders from './orders';
+import products from './products';
+import reviews from './reviews';
+import Segments from './segments/Segments';
+import visitors from './visitors';
+import { themes, ThemeName } from './themes/themes';
 
 const client = new Client();
 client
@@ -21,15 +34,63 @@ const dataProvider = appWriteDataProvider({
 });
 const authProvider = appWriteAuthProvider({ client });
 
-const App = () => (
-    <Admin
-        dataProvider={dataProvider}
-        authProvider={authProvider}
-        loginPage={LoginWithEmail}
-    >
-        <Resource name="contacts" {...contacts} />
-        <Resource name="companies" list={ListGuesser} edit={EditGuesser} />
-    </Admin>
+const i18nProvider = polyglotI18nProvider(
+    locale => {
+        if (locale === 'fr') {
+            return import('./i18n/fr').then(messages => messages.default);
+        }
+
+        // Always fallback on english
+        return englishMessages;
+    },
+    'en',
+    [
+        { locale: 'en', name: 'English' },
+        { locale: 'fr', name: 'Français' },
+    ]
 );
 
-export default App;
+const store = localStorageStore(undefined, 'ECommerce');
+
+const App = () => {
+    const [themeName] = useStore<ThemeName>('themeName', 'soft');
+    const singleTheme = themes.find(theme => theme.name === themeName)?.single;
+    const lightTheme = themes.find(theme => theme.name === themeName)?.light;
+    const darkTheme = themes.find(theme => theme.name === themeName)?.dark;
+    return (
+        <Admin
+            title="Posters Galore Admin"
+            dataProvider={dataProvider}
+            store={store}
+            authProvider={authProvider}
+            dashboard={Dashboard}
+            loginPage={Login}
+            layout={Layout}
+            i18nProvider={i18nProvider}
+            disableTelemetry
+            theme={singleTheme}
+            lightTheme={lightTheme}
+            darkTheme={darkTheme}
+            defaultTheme="light"
+            requireAuth
+        >
+            <CustomRoutes>
+                <Route path="/segments" element={<Segments />} />
+            </CustomRoutes>
+            <Resource name="customers" {...visitors} />
+            <Resource name="orders" {...orders} />
+            <Resource name="invoices" {...invoices} />
+            <Resource name="products" {...products} />
+            <Resource name="categories" {...categories} />
+            <Resource name="reviews" {...reviews} />
+        </Admin>
+    );
+};
+
+const AppWrapper = () => (
+    <StoreContextProvider value={store}>
+        <App />
+    </StoreContextProvider>
+);
+
+export default AppWrapper;
