@@ -1,10 +1,11 @@
-import React from 'react';
 import { createRoot } from 'react-dom/client';
+
 import App from './App';
 
-const root = createRoot(document.getElementById('root') as HTMLElement);
-root.render(
-    <React.StrictMode>
-        <App />
-    </React.StrictMode>
-);
+const container = document.getElementById('root');
+if (!container) {
+    throw new Error('No container found');
+}
+const root = createRoot(container);
+
+root.render(<App />);
